flesh out functions related to 'eb -a' support into new modules easybuild.tools.docs
