+++ easybuild/tools/docs.py
@@ -1,0 +1,168 @@
+# #
+# Copyright 2009-2014 Ghent University
+#
+# This file is part of EasyBuild,
+# originally created by the HPC team of Ghent University (http://ugent.be/hpc/en),
+# with support of Ghent University (http://ugent.be/hpc),
+# the Flemish Supercomputer Centre (VSC) (https://vscentrum.be/nl/en),
+# the Hercules foundation (http://www.herculesstichting.be/in_English)
+# and the Department of Economy, Science and Innovation (EWI) (http://www.ewi-vlaanderen.be/en).
+#
+# http://github.com/hpcugent/easybuild
+#
+# EasyBuild is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation v2.
+#
+# EasyBuild is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
+#
+# You should have received a copy of the GNU General Public License
+# along with EasyBuild.  If not, see <http://www.gnu.org/licenses/>.
+# #
+"""
+Documentation-related functionality
+
+@author: Stijn De Weirdt (Ghent University)
+@author: Dries Verdegem (Ghent University)
+@author: Kenneth Hoste (Ghent University)
+@author: Pieter De Baets (Ghent University)
+@author: Jens Timmerman (Ghent University)
+@author: Toon Willems (Ghent University)
+@author: Ward Poelmans (Ghent University)
+"""
+import copy
+
+from easybuild.framework.easyconfig.default import ALL_CATEGORIES, DEFAULT_CONFIG, HIDDEN, sorted_categories
+from easybuild.framework.easyconfig.easyconfig import get_easyblock_class
+from easybuild.tools.deprecated.eb_2_0 import ExtraOptionsDeprecatedReturnValue
+from easybuild.tools.ordereddict import OrderedDict
+from easybuild.tools.utilities import quote_str
+
+
+FORMAT_RST = 'rst'
+FORMAT_TXT = 'txt'
+
+
+def avail_easyconfig_params_rst(title, grouped_params):
+    """
+    Compose overview of available easyconfig parameters, in RST format.
+    """
+    def det_col_width(entries, title):
+        """Determine column width based on column title and list of entries."""
+        return max(map(len, entries + [title]))
+
+    # main title
+    lines = [
+        title,
+        '=' * len(title),
+        '',
+    ]
+
+    for grpname in grouped_params:
+        # group section title
+        lines.append("%s parameters" % grpname)
+        lines.extend(['-' * len(lines[-1]), ''])
+
+        name_title = "**Parameter name**"
+        descr_title = "**Description**"
+        dflt_title = "**Default value**"
+
+        # figure out column widths
+        nw = det_col_width(grouped_params[grpname].keys(), name_title) + 4  # +4 for raw format ("``foo``")
+        dw = det_col_width([x[0] for x in grouped_params[grpname].values()], descr_title)
+        dfw = det_col_width([str(quote_str(x[1])) for x in grouped_params[grpname].values()], dflt_title)
+
+        # 3 columns (name, description, default value), left-aligned, {c} is fill char
+        line_tmpl = "{0:{c}<%s}   {1:{c}<%s}   {2:{c}<%s}" % (nw, dw, dfw)
+        table_line = line_tmpl.format('', '', '', c='=', nw=nw, dw=dw, dfw=dfw)
+
+        # table header
+        lines.append(table_line)
+        lines.append(line_tmpl.format(name_title, descr_title, dflt_title, c=' '))
+        lines.append(line_tmpl.format('', '', '', c='-'))
+
+        # table rows by parameter
+        for name, (descr, dflt) in sorted(grouped_params[grpname].items()):
+            rawname = '``%s``' % name
+            lines.append(line_tmpl.format(rawname, descr, str(quote_str(dflt)), c=' '))
+        lines.append(table_line)
+        lines.append('')
+
+    return '\n'.join(lines)
+
+def avail_easyconfig_params_txt(title, grouped_params):
+    """
+    Compose overview of available easyconfig parameters, in plain text format.
+    """
+    # main title
+    lines = [
+        '%s:' % title,
+        '',
+    ]
+
+    for grpname in grouped_params:
+        # group section title
+        lines.append(grpname.upper())
+        lines.append('-' * len(lines[-1]))
+
+        # determine width of 'name' column, to left-align descriptions
+        nw = max(map(len, grouped_params[grpname].keys()))
+
+        # line by parameter
+        for name, (descr, dflt) in sorted(grouped_params[grpname].items()):
+            lines.append("{0:<{nw}}   {1:} [default: {2:}]".format(name, descr, str(quote_str(dflt)), nw=nw))
+        lines.append('')
+
+    return '\n'.join(lines)
+
+def avail_easyconfig_params(easyblock, output_format):
+    """
+    Compose overview of available easyconfig parameters, in specified format.
+    """
+    params = copy.deepcopy(DEFAULT_CONFIG)
+
+    # include list of extra parameters (if any)
+    extra_params = {}
+    app = get_easyblock_class(easyblock, default_fallback=False)
+    if app is not None:
+        extra_params = app.extra_options()
+    if isinstance(extra_params, ExtraOptionsDeprecatedReturnValue):
+        extra_params = dict(extra_params)
+    params.update(extra_params)
+
+    # compose title
+    title = "Available easyconfig parameters"
+    if extra_params:
+        title += " (* indicates specific to the %s easyblock)" % app.__name__
+
+    # group parameters by category
+    grouped_params = OrderedDict()
+    for category in sorted_categories():
+        # exclude hidden parameters
+        if category[1].upper() in [HIDDEN]:
+            continue
+
+        grpname = category[1]
+        grouped_params[grpname] = {}
+        for name, (dflt, descr, cat) in params.items():
+            # FIXME bug in default.py?
+            if isinstance(cat, basestring):
+                cat = ALL_CATEGORIES[cat]
+            if cat == category:
+                if name in extra_params:
+                    # mark easyblock-specific parameters
+                    name = '%s*' % name
+                grouped_params[grpname].update({name: (descr, dflt)})
+
+        if not grouped_params[grpname]:
+            del grouped_params[grpname]
+
+    # compose output, according to specified format (txt, rst, ...)
+    avail_easyconfig_params_functions = {
+        FORMAT_RST: avail_easyconfig_params_rst,
+        FORMAT_TXT: avail_easyconfig_params_txt,
+    }
+    return avail_easyconfig_params_functions[output_format](title, grouped_params)
